+++ src/app/page.tsx
@@ -1,8 +1,7 @@
 'use client'
 
 import { Suspense } from 'react'
-import { useSearchParams } from 'next/navigation'
-import { useEffect } from 'react'
+import ScrollToHero from '@/components/ScrollToHero'
 import HeroSection from '@/components/HeroSection'
 import PopularServices from '@/components/PopularServices'
 import AboutSection from '@/components/AboutSection'
@@ -10,29 +9,9 @@ import WhyChooseUsSection from '@/components/WhyChooseUsSection'
 import HowItWorksSection from '@/components/HowItWorksSection'
 
 function HomePageContent() {
-  const searchParams = useSearchParams()
-
-  useEffect(() => {
-    const shouldScroll = searchParams.get('scrollToHero') === 'true'
-    if (shouldScroll) {
-      const timeout = setTimeout(() => {
-        const target = document.getElementById('herosection')
-        if (target) {
-          target.scrollIntoView({ behavior: 'smooth' })
-
-          // ลบ query ออกเพื่อไม่ scroll ซ้ำอีก
-          const url = new URL(window.location.href)
-          url.searchParams.delete('scrollToHero')
-          window.history.replaceState({}, '', url.toString())
-        }
-      }, 50)
-
-      return () => clearTimeout(timeout)
-    }
-  }, [searchParams])
-
   return (
     <>
+      <ScrollToHero />
       <HeroSection />
       <PopularServices />
       <AboutSection />
